+++ applications/fishing-map/src/data/config.ts
@@ -10,6 +10,8 @@ export const IS_PRODUCTION =
 export const API_GATEWAY = process.env.API_GATEWAY || process.env.REACT_APP_API_GATEWAY || ''
 export const CARRIER_PORTAL_URL =
   process.env.REACT_APP_CARRIER_PORTAL_URL || 'https://carrier-portal.dev.globalfishingwatch.org'
+export const LATEST_CARRIER_DATASET_ID =
+  process.env.REACT_APP_LATEST_CARRIER_DATASET_ID || 'carriers:latest'
 
 export const GOOGLE_UNIVERSAL_ANALYTICS_ID = process.env.REACT_APP_GOOGLE_UNIVERSAL_ANALYTICS_ID
 export const GOOGLE_UNIVERSAL_ANALYTICS_INIT_OPTIONS: ReactGA.InitializeOptions = IS_PRODUCTION
@@ -24,7 +26,7 @@ export const FULL_SUFIX = 'full'
 export const PRIVATE_SUFIX = 'private'
 
 // used when no url data and no workspace data
-const end = DateTime.fromObject({ hour: 0, minute: 0, second: 0, zone: 'utc' })
+export const LAST_DATA_UPDATE = DateTime.fromObject({ hour: 0, minute: 0, second: 0, zone: 'utc' })
   .minus({ days: 3 })
   .toISO()
 
@@ -34,8 +36,8 @@ export const DEFAULT_VIEWPORT = {
   zoom: 1,
 }
 export const DEFAULT_TIME_RANGE = {
-  start: DateTime.fromISO(end).minus({ months: 3 }).toISO(),
-  end,
+  start: DateTime.fromISO(LAST_DATA_UPDATE).minus({ months: 3 }).toISO(),
+  end: LAST_DATA_UPDATE,
 }
 
 export const DEFAULT_ACTIVITY_CATEGORY = 'fishing'