+++ packages/layer-composer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @globalfishingwatch/layer-composer
 
+## 8.6.0
+
+### Minor Changes
+
+- 3513a689: support interaction (configured in dataview as presenceDetails) for presence layers
+
 ## 8.5.0
 
 ### Minor Changes